# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GL Social API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get public user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/me/verify-age:
    post:
      operationId: verifyAge
      tags: [users]
      summary: Submit age verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgeVerification"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Posts ───────────────────────────────────────────────────────────────────
...
  /raffles:
    get:
      operationId: listRaffles
      tags: [raffles]
      summary: List raffles
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [active, upcoming, ended, null]
      responses:
        "200":
          description: Raffles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Raffle"

    post:
      operationId: createRaffle
      tags: [raffles]
      summary: Create a raffle (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaffleInput"
      responses:
        "201":
          description: Created raffle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Raffle"
  /raffles/{raffleId}:
    get:
      operationId: getRaffle
      tags: [raffles]
      summary: Get a raffle by ID
      parameters:
        - name: raffleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Raffle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Raffle"
        "404":
          description: Not found
  /raffles/{raffleId}/enter:
    post:
      operationId: enterRaffle
      tags: [raffles]
      summary: Enter a raffle using GL Coins
      parameters:
        - name: raffleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaffleEntry"
        "400":
          description: Insufficient coins or already entered
  /raffles/{raffleId}/draw:
    post:
      operationId: drawRaffleWinner
      tags: [raffles]
      summary: Draw raffle winner (admin only)
      parameters:
        - name: raffleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Winner info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaffleWinner"
  /raffles/{raffleId}/entries:
    get:
      operationId: listRaffleEntries
      tags: [raffles]
      summary: List entries for a raffle (admin only)
      parameters:
        - name: raffleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RaffleEntry"

  # ─── Follows ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, role, coinBalance, isVerified, isBanned, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [free, premium, admin]
        coinBalance:
          type: integer
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        ageVerified:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
...
    AgeVerification:
      type: object
      required: [birthYear]
      properties:
        birthYear:
          type: integer
...
    Raffle:
      type: object
      required: [id, title, description, entryCost, startTime, endTime, status, entryCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        prize:
          type: ["string", "null"]
        entryCost:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [upcoming, active, ended]
        entryCount:
          type: integer
        winnerId:
          type: ["integer", "null"]
        winner:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        hasEntered:
          type: boolean
        createdAt:
          type: string
    RaffleInput:
      type: object
      required: [title, description, entryCost, startTime, endTime]
      properties:
        title:
          type: string
        description:
          type: string
        prize:
          type: ["string", "null"]
        entryCost:
          type: integer
          minimum: 1
        startTime:
          type: string
        endTime:
          type: string
    RaffleEntry:
      type: object
      required: [id, raffleId, userId, user, createdAt]
      properties:
        id:
          type: integer
        raffleId:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
    RaffleWinner:
      type: object
      required: [raffleId, winner]
      properties:
        raffleId:
          type: integer
        winner:
          $ref: "#/components/schemas/User"
...